Move StateMerging to runners/core-java
