+++ runners/core-java/src/main/java/org/apache/beam/runners/core/SystemReduceFn.java
@@ -31,7 +31,6 @@ import org.apache.beam.sdk.util.state.CombiningState;
 import org.apache.beam.sdk.util.state.MergingStateAccessor;
 import org.apache.beam.sdk.util.state.ReadableState;
 import org.apache.beam.sdk.util.state.StateAccessor;
-import org.apache.beam.sdk.util.state.StateMerging;
 import org.apache.beam.sdk.util.state.StateTag;
 import org.apache.beam.sdk.util.state.StateTags;
